+++ .phan/plugins/UnusedSuppressionPlugin.php
@@ -1,8 +1,6 @@
 <?php declare(strict_types=1);
 
-use Phan\AST\AnalysisVisitor;
 use Phan\CodeBase;
-use Phan\Language\Context;
 use Phan\Language\Element\AddressableElement;
 use Phan\Language\Element\Clazz;
 use Phan\Language\Element\Func;
@@ -14,7 +12,6 @@ use Phan\PluginV2\AnalyzeFunctionCapability;
 use Phan\PluginV2\AnalyzePropertyCapability;
 use Phan\PluginV2\AnalyzeMethodCapability;
 use Phan\PluginV2\FinalizeProcessCapability;
-use ast\Node;
 
 /**
  * Check for unused (at)suppress annotations.