+++ src/Phan/Daemon.php
@@ -2,7 +2,6 @@
 namespace Phan;
 
 use Phan\Daemon\Request;
-use Phan\Library\FileCache;
 
 /**
  * an analyzing daemon, to be used by IDEs.