+++ src/Phan/Language/Element/Comment/Parameter.php
@@ -4,7 +4,6 @@ namespace Phan\Language\Element\Comment;
 
 use Phan\Language\Context;
 use Phan\Language\Element\Variable;
-use Phan\Language\Type\NullType;
 use Phan\Language\UnionType;
 
 class Parameter