+++ src/Phan/LanguageServer/ProtocolStreamReader.php
@@ -8,6 +8,7 @@ use Phan\LanguageServer\Protocol\Message;
 use AdvancedJsonRpc\Message as MessageBody;
 use Sabre\Event\Loop;
 use Sabre\Event\Emitter;
+
 use Exception;
 
 /**
@@ -99,7 +100,7 @@ class ProtocolStreamReader extends Emitter implements ProtocolReader
                         // MessageBody::parse can throw an Error, maybe log an error?
                         try {
                             $msg = new Message(MessageBody::parse($this->buffer), $this->headers);
-                        } catch (\Exception $e) {
+                        } catch (Exception $e) {
                             $msg = null;
                         }
                         if ($msg) {