+++ src/Phan/Plugin/Internal/ArrayReturnTypeOverridePlugin.php
@@ -5,13 +5,9 @@ use Phan\CodeBase;
 use Phan\Analysis\ArgumentType;
 use Phan\Analysis\PostOrderAnalysisVisitor;
 use Phan\AST\UnionTypeVisitor;
-use Phan\Issue;
 use Phan\Config;
 use Phan\Language\Context;
 use Phan\Language\Element\Func;
-use Phan\Language\Element\FunctionInterface;
-use Phan\Language\FQSEN\FullyQualifiedMethodName;
-use Phan\Language\FQSEN\FullyQualifiedFunctionName;
 use Phan\Language\Type\ArrayType;
 use Phan\Language\Type\FalseType;
 use Phan\Language\Type\IntType;
@@ -20,9 +16,7 @@ use Phan\Language\Type\MixedType;
 use Phan\Language\Type\StringType;
 use Phan\Language\UnionType;
 use Phan\PluginV2\ReturnTypeOverrideCapability;
-use Phan\PluginV2\AnalyzeFunctionCallCapability;
 use Phan\PluginV2;
-use ast\Node;
 
 /**
  * NOTE: This is automatically loaded by phan. Do not include it in a config.